+++ ChangeLog.md
@@ -2,6 +2,8 @@
 
 ## 0.6.3 (2026-05-27)
 - fix 0.6.2 regression in dnf detection logic that defaulted to /usr/bin/dnf5
+- with dnf5 `--save` no longer uses config-manager since
+  `/etc/dnf/repos.override.d` is not supported yet
 
 ## 0.6.2 (2026-03-15)
 - `--save` now works without dnf using sed